Making dev.gitlive auth dependency private to auth module
