+++ auth/src/androidMain/kotlin/com/leandrodev/auth/signout/SignOutView.kt
@@ -1,0 +1,30 @@
+package com.leandrodev.auth.signout
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import dev.gitlive.firebase.Firebase
+import dev.gitlive.firebase.auth.auth
+import kotlinx.coroutines.launch
+
+@Composable
+fun SignOutView(
+    modifier: Modifier = Modifier
+) {
+    val scope = rememberCoroutineScope()
+    Column(modifier) {
+        Text(text = "Authenticated.", modifier = Modifier.padding(top = 32.dp))
+        Button(onClick = {
+            scope.launch {
+                Firebase.auth.signOut()
+            }
+        }) {
+            Text(text = "Sign out")
+        }
+    }
+}
