+++ bills/src/androidMain/kotlin/com/leandrodev/bills/wallet/WalletListScreen.kt
@@ -3,32 +3,27 @@ package com.leandrodev.bills.wallet
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.Button
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.leandrodev.auth.signout.SignOutView
 import com.leandrodev.bills.wallet.list.WalletListState
 import com.leandrodev.bills.wallet.list.WalletListViewModel
 import com.leandrodev.bills.wallet.model.Wallet
-import com.leandrodev.ui.theme.HomeTheme
 import com.leandrodev.shared.MutableViewState
 import com.leandrodev.shared.createViewAction
 import com.leandrodev.shared.getViewModel
-import dev.gitlive.firebase.Firebase
-import dev.gitlive.firebase.auth.auth
-import kotlinx.coroutines.launch
+import com.leandrodev.ui.theme.HomeTheme
 
 @Composable
 fun WalletListScreen(
     modifier: Modifier = Modifier,
     viewModel: WalletListViewModel = getViewModel(),
 ) {
-    val scope = rememberCoroutineScope()
     val state: WalletListState by viewModel.state.collectAsState(initial = WalletListState.Loading)
 
     Column(modifier) {
@@ -52,14 +47,9 @@ fun WalletListScreen(
                 }
             }
         }
-        Text(text = "Authenticated.", modifier = Modifier.padding(top = 32.dp))
-        Button(onClick = {
-            scope.launch {
-                Firebase.auth.signOut()
-            }
-        }) {
-            Text(text = "Sign out")
-        }
+        SignOutView(
+            modifier = Modifier.padding(16.dp)
+        )
     }
 }
